Consolidate MicroProfile FAT tests into one project and add some MP Compatible tests
