--- dev/com.ibm.ws.microprofile.1.0_fat/fat/src/com/ibm/ws/microprofile10/fat/tests/helloworld/HelloWorldResource.java
@@ -1,35 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2019 IBM Corporation and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *
- * Contributors:
- *     IBM Corporation - initial API and implementation
- *******************************************************************************/
-package com.ibm.ws.microprofile10.fat.tests.helloworld;
-
-import javax.inject.Inject;
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
-
-@Path("/helloworld")
-public class HelloWorldResource {
-
-    @Inject
-    HelloWorldBean bean;
-
-    /**
-     * Processes a GET request and returns the stored message.
-     *
-     * @return the stored message
-     */
-    @GET
-    @Produces(MediaType.TEXT_PLAIN)
-    public String getMessage() {
-        return bean.getMessage();
-    }
-}
