--- dev/com.ibm.ws.microprofile.1.2_fat/.settings/bndtools.core.prefs
@@ -1,2 +1,0 @@
-compileErrorAction=build
-eclipse.preferences.version=1
