--- dev/com.ibm.ws.microprofile.1.2_fat/.settings/org.eclipse.core.resources.prefs
@@ -1,3 +1,0 @@
-#Wed Aug 17 16:19:12 BST 2011
-eclipse.preferences.version=1
-encoding/<project>=UTF-8
