--- dev/com.ibm.ws.microprofile.1.2_fat/bnd.bnd
@@ -1,27 +1,0 @@
-#*******************************************************************************
-# Copyright (c) 2019 IBM Corporation and others.
-# All rights reserved. This program and the accompanying materials
-# are made available under the terms of the Eclipse Public License v1.0
-# which accompanies this distribution, and is available at
-# http://www.eclipse.org/legal/epl-v10.html
-#
-# Contributors:
-#     IBM Corporation - initial API and implementation
-#*******************************************************************************
--include= ~../cnf/resources/bnd/bundle.props
-bVersion=1.0
-
-src: \
-	fat/src
-
-	
-fat.project: true
-
-tested.features: microProfile-1.2, microProfile-1.3, microProfile-1.4, microProfile-2.0, microProfile-2.1, microProfile-2.2, microProfile-3.0, microProfile-3.2, microProfile-3.3, microProfile-4.0
-
--buildpath: \
-	com.ibm.websphere.javaee.servlet.3.1;version=latest,\
-	com.ibm.websphere.javaee.jaxrs.2.0;version=latest,\
-        com.ibm.websphere.javaee.annotation.1.2;version=latest,\
-	com.ibm.websphere.javaee.cdi.1.2;version=latest,\
-	com.ibm.websphere.org.eclipse.microprofile.config.1.1;version=latest
