--- dev/com.ibm.ws.microprofile.1.2_fat/fat/src/com/ibm/ws/microprofile12/fat/suite/FATSuite.java
@@ -1,35 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2019 IBM Corporation and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *
- * Contributors:
- *     IBM Corporation - initial API and implementation
- *******************************************************************************/
-package com.ibm.ws.microprofile12.fat.suite;
-
-import org.junit.BeforeClass;
-import org.junit.runner.RunWith;
-import org.junit.runners.Suite;
-import org.junit.runners.Suite.SuiteClasses;
-
-import com.ibm.ws.fat.util.FatLogHandler;
-
-@RunWith(Suite.class)
-@SuiteClasses({
-                SimpleJAXRSCDITest.class,
-})
-
-public class FATSuite {
-
-    /**
-     * @see {@link FatLogHandler#generateHelpFile()}
-     */
-    @BeforeClass
-    public static void generateHelpFile() {
-        FatLogHandler.generateHelpFile();
-    }
-
-}
