+++ dev/io.openliberty.microprofile.internal_fat/.classpath
