+++ dev/io.openliberty.microprofile.internal_fat/.settings/bndtools.core.prefs
