+++ dev/io.openliberty.microprofile.internal_fat/publish/servers/MPCompatibleServer/MP10.xml
