+++ dev/io.openliberty.microprofile.internal_fat/publish/servers/MPCompatibleServer/MP12.xml
