+++ dev/io.openliberty.microprofile.internal_fat/publish/servers/MPCompatibleServer/MP33plusStandalone.xml
@@ -1,0 +1,10 @@
+<server description="Server for testing microProfile">
+
+    <include location="../fatTestPorts.xml"/>
+
+    <featureManager>
+        <feature>microProfile-3.3</feature>
+        <feature>mpGraphQL-1.0</feature>
+        <feature>mpReactiveMessaging-1.0</feature>
+    </featureManager>
+</server>
