+++ dev/io.openliberty.microprofile.internal_fat/publish/servers/MPCompatibleServer/MP40.xml
@@ -1,0 +1,8 @@
+<server description="Server for testing microProfile">
+
+    <include location="../fatTestPorts.xml"/>
+
+    <featureManager>
+        <feature>microProfile-4.0</feature>
+    </featureManager>
+</server>
