+++ dev/io.openliberty.microprofile.internal_fat/publish/servers/MPCompatibleServer/MP40andMPConfig14.xml
@@ -1,0 +1,9 @@
+<server description="Server for testing microProfile">
+
+    <include location="../fatTestPorts.xml"/>
+
+    <featureManager>
+        <feature>microProfile-4.0</feature>
+        <feature>mpConfig-1.4</feature>
+    </featureManager>
+</server>
