+++ dev/io.openliberty.microprofile.internal_fat/publish/servers/MPCompatibleServer/MP40plusStandalone.xml
@@ -1,0 +1,11 @@
+<server description="Server for testing microProfile">
+
+    <include location="../fatTestPorts.xml"/>
+
+    <featureManager>
+        <feature>microProfile-4.0</feature>
+        <feature>mpGraphQL-1.0</feature>
+        <!-- TODO https://github.com/OpenLiberty/open-liberty/issues/15440 -->
+        <!-- <feature>mpReactiveMessaging-1.0</feature>-->
+    </featureManager>
+</server>
