+++ dev/io.openliberty.microprofile.internal_fat/publish/servers/MPCompatibleServer/bootstrap.properties
