+++ dev/io.openliberty.microprofile.internal_fat/publish/servers/MPServer/bootstrap.properties
